created a types module to fix circular reference error
